backend/requirements.txt
Django==5.0.7
psycopg==3.2.1
python-dotenv==1.0.1
djangorestframework==3.15.2
django-filter==24.2
